Seeds (#5)
allow setting a random seed from the config or even the main args (args more powerful) to reproduce previous experiments. if no random seed given in config and args, generates a random one and uses that. sets the final choice of seed in train_params to allow future reproductions if wanted